[redis] Patch on module import
