+++ ddtrace/contrib/redis/tracers.py
@@ -1,5 +1,3 @@
-from redis import StrictRedis
-
 from ...utils.deprecation import deprecated
 
 
@@ -8,6 +6,7 @@ DEFAULT_SERVICE = 'redis'
 
 @deprecated(message='Use patching instead (see the docs).', version='1.0.0')
 def get_traced_redis(ddtracer, service=DEFAULT_SERVICE, meta=None):
+    from redis import StrictRedis
     return _get_traced_redis(ddtracer, StrictRedis, service, meta)
